refactor(providers): promote TurnTrace to providers/shared
anthropic-direct/loop/turn-trace.ts's TurnTrace class and
openai-compatible/query.ts's inline equivalent both bracket a provider
turn with loop_start / interrupt_halt / loop_end witness events plus
the interrupt-timestamp bookkeeping the halt-latency measurement
needs -- openai-compatible's own comments already described itself as
mirroring the anthropic-direct version. ~70% overlapping logic.

Promoted the class to providers/shared/turn-trace.ts. The only real
behavioral parameter was the `provider` label stamped onto the
interrupt_halt metadata (each side hardcoded its own literal) -- added
as an explicit third constructor argument so each provider still tags
its own halt-latency events correctly; nothing else changed.

Verified before promoting: read both originals in full side by side.
The class already has zero Anthropic-SDK-specific coupling. The one
apparent ordering difference (openai-compatible set up the interrupt
listener BEFORE emitting loop_start; anthropic-direct emits loop_start
THEN sets up the listener) has no observable effect -- both statements
are synchronous with no await between them, so no abort can land in
between regardless of order; the class's own doc comment already notes
this ordering is about trace-stream byte-comparability with historical
sessions, not correctness.

Verification: full suite green before and after, both under the
default (parallel) and --no-file-parallelism (serial) pools, matching
the clean baseline exactly: 773 test files / 14746 passed / 2 skipped,
0 failed.

Note: a sibling finding in the same investigation (unifying
anthropic-direct/loop/tool-dispatch.ts and openai-compatible/query/
dispatch-append.ts's try/catch skeleton) was attempted and REVERTED --
it reproduced a full-suite-only heap-OOM crash traced to adopting a
more "defensive" try/catch boundary shape that turned out to have a
real behavioral dependency somewhere in the surrounding turn/retry
machinery. Not included in this commit; needs its own investigation
before attempting again.
